+++ PersistencePlugin/src/com/elmakers/mine/bukkit/plugins/persistence/data/DataField.java
@@ -1,5 +1,10 @@
 package com.elmakers.mine.bukkit.plugins.persistence.data;
 
+import java.util.Date;
+import java.util.logging.Logger;
+
+import com.elmakers.mine.bukkit.plugins.persistence.Persistence;
+
 /**
  * Represents a single field in a data row.
  * 
@@ -49,6 +54,69 @@ public class DataField
 		return value;
 	}
 	
+	public Object getValue(Class<?> targetClass)
+	{
+		if (value == null) 
+		{
+			return null;
+		}
+		
+		Class<?> valueClass = value.getClass();
+		
+		if (targetClass.isAssignableFrom(valueClass))
+		{
+			return value;
+		}
+		
+		if (targetClass.isEnum() && (int.class.isAssignableFrom(valueClass) || Integer.class.isAssignableFrom(valueClass)))
+		{
+			int ordinal = (Integer)value;
+			return targetClass.getEnumConstants()[ordinal];
+		}
+		
+		if (Date.class.isAssignableFrom(targetClass))
+		{
+			if (Integer.class.isAssignableFrom(valueClass) || int.class.isAssignableFrom(valueClass))
+			{
+				Integer intDate = (Integer)value;
+				Date d = new Date(intDate * 1000);
+				return d;
+			}
+		}
+		
+		if (boolean.class.isAssignableFrom(targetClass) || Boolean.class.isAssignableFrom(targetClass))
+		{
+			if (Boolean.class.isAssignableFrom(valueClass) || boolean.class.isAssignableFrom(valueClass))
+			{
+				return targetClass.cast((Boolean)value);	
+			}
+			if (Integer.class.isAssignableFrom(valueClass) || int.class.isAssignableFrom(valueClass))
+			{
+				Integer intBoolean = (Integer)value;
+				Boolean b = intBoolean != 0;
+				return b;	
+			}
+		}
+		
+		if (float.class.isAssignableFrom(targetClass) || Float.class.isAssignableFrom(targetClass))
+		{
+			if (Float.class.isAssignableFrom(valueClass) || float.class.isAssignableFrom(valueClass))
+			{
+				return targetClass.cast((Float)value);	
+			}
+			if (Integer.class.isAssignableFrom(valueClass) || int.class.isAssignableFrom(valueClass))
+			{
+				return (Float)(float)(Integer)value;	
+			}
+			if (Double.class.isAssignableFrom(valueClass) || double.class.isAssignableFrom(valueClass))
+			{
+				return (Float)(float)(double)(Double)value;	
+			}
+		}
+		
+		return value;
+	}
+	
 	public void setName(String name)
 	{
 		this.name = name;
@@ -78,9 +146,10 @@ public class DataField
 		return idField;
 	}
 
-	private DataType 	type;
-	private Object		value;
-	private String		name;
-	private boolean		autogenerate;
-	private boolean		idField;
+	protected static Logger log = Persistence.getLogger();
+	protected DataType 		type;
+	protected Object		value;
+	protected String		name;
+	protected boolean		autogenerate;
+	protected boolean		idField;
 }